+++ include/core/intersect-info.hpp
@@ -1,6 +1,6 @@
 #ifndef _INTERSECT_INFO_H
 #define _INTERSECT_INFO_H
-#include "vec3.hpp"
+#include "core/vec3.hpp"
 
 struct IntersectInfo {
   float t;
