+++ include/core/vec3.hpp
